backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# AI and ML
anthropic==0.7.8
openai==1.6.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1

# Task Queue and Messaging
celery==5.3.4
redis==5.0.1

# File Storage
minio==7.2.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Vector Database
qdrant-client==1.7.0

# Time Series Database
influxdb-client==1.39.0

# Template Engine
jinja2==3.1.2

# Image Processing
pillow==10.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1